modify structure again after meeting
