+++ src/main/java/com/greenfox/poker/service/UserService.java
@@ -1,7 +1,6 @@
 package com.greenfox.poker.service;
 
-import com.greenfox.poker.model.ResponseDataObject;
-import com.greenfox.poker.model.User;
+import com.greenfox.poker.model.PokerUser;
 import com.greenfox.poker.model.UserRegister;
 import org.springframework.stereotype.Component;
 
@@ -11,8 +10,8 @@ public class UserService {
   public UserService() {
   }
 
-  public ResponseDataObject registerNewUser(UserRegister newUserRegister){
-    User newUser = new User(newUserRegister.getUsername(), newUserRegister.getPassword(),newUserRegister.getEmail(),newUserRegister.getAvatar());
-    return new ResponseDataObject(newUser);
+  public PokerUser registerNewUser(PokerUser newUserToRegister){
+    PokerUser pokerUser = new PokerUser(newUserToRegister.getUsername(), newUserToRegister.getPassword(),newUserToRegister.getEmail(),newUserToRegister.getAvatar());
+    return pokerUser;
   }
 }
